requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
terminaltables==3.1.0
urllib3==1.24.1
